Remove redundant `ConfigSyncMode` type
